refactor(sx): autostart redshift as systemd service
dex requires x-terminal-emulator to autostart terminal based desktop
files. The workaround is to start redshift as a user service in sxrc.

A cleaner solution would be to enable the redshift service in the
graphical target and not start it from sxrc, but we currently boot to
multi-user.target and start xorg from the login shell.